+++ src/Resources/lang/nl/app.php
@@ -2,6 +2,15 @@
 
 return [
     'admin' => [
+        'common' => [
+            'error' => [
+                'resource-not-found' => 'Bron niet gevonden.',
+                'empty-payload' => 'Er zijn geen producten verzonden.',
+                'not-a-product-list' => 'De aanvraagbody moet een lijst met producten zijn.',
+                'not-a-product-object' => 'Dit item is geen productobject.',
+            ],
+        ],
+
         'sales' => [
             'orders' => [
                 'cancel-success' => 'Bestelling succesvol geannuleerd.',